+++ apps/router-e2e/__e2e__/server-inserted-html/app/_layout.tsx
@@ -1,0 +1,11 @@
+import { Slot } from 'expo-router';
+
+import { StreamedDataProvider } from '../components/StreamedData';
+
+export default function RootLayout() {
+  return (
+    <StreamedDataProvider>
+      <Slot />
+    </StreamedDataProvider>
+  );
+}