+++ packages/expo-router/src/exports.ts
@@ -30,6 +30,12 @@ export { ErrorBoundary } from './views/ErrorBoundary';
 export { SuspenseFallback, type SuspenseFallbackProps } from './views/SuspenseFallback';
 export type { ScreenProps } from './useScreens';
 
+// Server rendering
+export {
+  useServerInsertedHTML,
+  type ServerInsertedHTMLCallback,
+} from './server/ServerInsertedHTML';
+
 // Platform
 /**
  * @hidden