+++ packages/expo-router/src/internal/server.ts
@@ -1,1 +1,2 @@
 export * from '../server/ServerDocument';
+export * from '../server/ServerInsertedHTML';